requirements.txt
py-clob-client>=0.15.0
web3>=6.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
apscheduler>=3.10.0
openai>=1.30.0
scipy>=1.12.0
aiosqlite>=0.20.0
loguru>=0.7.0
cryptography>=42.0.0
websockets>=12.0